Rename version.h to leaf_version.h
