+++ src/vario/.gitignore
@@ -1,2 +1,2 @@
-version.h.lastbuild
-version.h.original
+leaf_version.h.lastbuild
+leaf_version.h.original
